Extract Canvas to own file
